refactor(mon-pix): move password reset acceptance tests in authentication
